refactor: extract shared utilities and add trailer key constant
